// src/shared/dataprovider/api/model/transactionSummary.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionSummaryData } from './transactionSummaryData';
import { TransactionSummaryDataTotal } from './transactionSummaryDataTotal';
import { TransactionSummaryIncomeTotal } from './transactionSummaryIncomeTotal';

export interface TransactionSummary { 
    TotalRecordCount?: number;
    PageNumber?: number;
    PageSize?: number;
    Total?: TransactionSummaryDataTotal;
    TotalIncome?: TransactionSummaryIncomeTotal;
    Data?: Array<TransactionSummaryData>;
}