// src/shared/dataprovider/api/model/clientWalletDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProviderDTO } from './providerDTO';
import { StatusEnum } from './statusEnum';

export interface ClientWalletDTO { 
    Id?: number;
    Name?: string;
    TransactionCount?: number;
    SettlementWalletBalance?: number;
    TotalSettlementWalletWithdrawn?: number;
    TotalCashIn?: number;
    TotalCashOut?: number;
    AvailableBalance?: number;
    Status?: StatusEnum;
    ProviderId?: number;
    StartDate?: Date;
    EndDate?: Date;
    FilteredTotalWalletWithdrawn?: number;
    FilteredTotalCashIn?: number;
    FilteredTotalCashOut?: number;
    Provider?: ProviderDTO;
}