// src/shared/dataprovider/api/model/clientWalletListDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ClientWalletDTO } from './clientWalletDTO';

export interface ClientWalletListDTO { 
    TotalRecordCount?: number;
    PageSize?: number;
    PageNumber?: number;
    Data?: Array<ClientWalletDTO>;
}